Modify archetype to enable checkstyle failOnError by default
As a community, we want to try our best to influence new projects to write
clean, consistent code.  As such, we really ought to enable checkstyle failOnError
for projects that are generated using the startup archetype.  This change does
that;  projects can still turn this off, but it requires manual opt out.

By default, we should really try to influence failOnError so people write
good looking code.

Change-Id: If740df92cbc3d8ff8a0169d323028415e239a15a
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>